Added option -min-level=error|warning|info|hint|unused|maybe|syntax to filter issues by severity Also did refactoring: - moved flags to flags.go (but global variables are evil) - added new type SeverityLevel same as int
